добавил репозитории

+++ src/main/java/org/example/secretsanta/repository/InviteRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InviteRepository extends JpaRepository<InviteEntity, Long> {
+public interface InviteRepository extends JpaRepository<InviteEntity, Integer> {
 }
