+++ src/main/java/org/example/secretsanta/repository/ResultRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ResultRepository extends JpaRepository<ResultEntity, Long> {
+public interface ResultRepository extends JpaRepository<ResultEntity, Integer> {
 }
