+++ src/main/java/org/example/secretsanta/repository/UserInfoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserInfoRepository extends JpaRepository<UserInfoEntity, Long> {
-    
+public interface UserInfoRepository extends JpaRepository<UserInfoEntity, Integer> {
+
 }
