+++ src/main/java/org/example/secretsanta/repository/WishRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WishRepository extends JpaRepository<WishEntity, Long> {
+public interface WishRepository extends JpaRepository<WishEntity, Integer> {
 }
